requirements.txt
altair==5.5.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
flatbuffers==25.1.24
gitdb==4.0.12
GitPython==3.1.44
humanfriendly==10.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy_loader==0.4
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.23.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
onnxruntime==1.20.1
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
protobuf==5.29.3
pyarrow==19.0.0
pydeck==0.9.1
Pygments==2.19.1
PyMatting==1.1.13
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytz==2024.2
referencing==0.36.2
rembg==2.0.61
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
scikit-image==0.25.0
scipy==1.15.1
six==1.17.0
smmap==5.0.2
streamlit==1.41.1
sympy==1.13.3
tenacity==9.0.0
tifffile==2025.1.10
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
